+++ crates/miden-tx/src/executor/exec_host.rs
@@ -9,6 +9,7 @@ use miden_processor::{BaseHost, FutureMaybeSend, Host, LoadedMastForest, Process
 use miden_protocol::account::auth::PublicKeyCommitment;
 use miden_protocol::account::{
     AccountCode,
+    AccountCodeUpgrade,
     AccountId,
     AccountPatch,
     PartialAccount,
@@ -133,7 +134,8 @@ where
         ref_block: BlockNumber,
         block_commitments: BTreeMap<BlockNumber, Word>,
         source_manager: Arc<dyn SourceManagerSync>,
-    ) -> Self {
+        account_code_upgrade: Option<AccountCodeUpgrade>,
+    ) -> Result<Self, TransactionKernelError> {
         let base_host = TransactionBaseHost::new(
             account,
             input_notes,
@@ -141,9 +143,10 @@ where
             mast_store,
             scripts_mast_store,
             acct_procedure_index_map,
-        );
+            account_code_upgrade,
+        )?;
 
-        Self {
+        Ok(Self {
             base_host,
             tx_progress: TransactionProgress::default(),
             authenticator,
@@ -153,7 +156,7 @@ where
             generated_signatures: BTreeMap::new(),
             in_auth_procedure: false,
             source_manager,
-        }
+        })
     }
 
     // PUBLIC ACCESSORS
@@ -553,6 +556,10 @@ where
 
                 TransactionEvent::AccountPushProcedureIndex { code_commitment, procedure_root } => {
                     self.base_host.on_account_push_procedure_index(code_commitment, procedure_root)
+                },
+
+                TransactionEvent::AccountUpgradeInitialized { new_code_commitment } => {
+                    self.base_host.on_account_upgrade_initialized(new_code_commitment)
                 },
 
                 TransactionEvent::NoteBeforeCreated { note_idx, metadata, recipient_data } => {