+++ crates/miden-tx/src/prover/mod.rs
@@ -155,7 +155,7 @@ impl LocalTransactionProver {
 
         let block_commitments = tx_inputs.collect_block_commitments();
 
-        let (partial_account, ref_block, _, input_notes, _) = tx_inputs.into_parts();
+        let (partial_account, ref_block, _, input_notes, tx_args) = tx_inputs.into_parts();
         let mut host = TransactionProverHost::new(
             &partial_account,
             input_notes,
@@ -163,7 +163,9 @@ impl LocalTransactionProver {
             &mast_store,
             script_mast_store,
             account_procedure_index_map,
-        );
+            tx_args.account_code_upgrade().cloned(),
+        )
+        .map_err(TransactionProverError::TransactionHostConstructionFailed)?;
 
         let advice_inputs = advice_inputs.into_advice_inputs();